refactor: OAuth クライアント生成を service へ移す
handlers/auth.rs の private ロジックを services/auth.rs へ移し、handler は service 呼び出しのみに整理。\n\nCo-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
